perf(runner): migrate runner opt logs to examples/bench
