+++ src/lib.rs
@@ -62,12 +62,12 @@ extern crate log;
 extern crate serde;
 
 mod conf;
-mod result;
 mod shared;
 
 pub mod api;
 pub mod chain;
 pub mod cmd;
 pub mod mmr;
+pub mod result;
 
 pub use shared::ShadowShared;
